FIX: Separate Drogon .fmu function to own module
